solidity upgrade

+++ contracts/contracts/IMACI.sol
@@ -5,7 +5,7 @@ import { AccQueue } from "./trees/AccQueue.sol";
 
 interface IMACI {
 
-    function stateTreeDepth() external view returns (uint8);
+    function stateTreeDepth() external view returns (uint256);
     function vkRegistry() external view returns (VkRegistry);
     function getStateAqRoot() external view returns (uint256);
     function mergeStateAqSubRoots(uint256 _numSrQueueOps, uint256 _pollId) external;
